refactor: convert cell renderer utilities to hooks for context integration
Previous getCellRenderer() and cellToString() functions blocked React
Context integration needed for the custom cell renderer plugin system.
Functions cannot access React Context, preventing injection of custom
renderers from CellProvider as outlined in the design document.

This change converts both utilities to hooks that return factory functions:
- useGetCellRenderer() returns (props) => CellRenderer
- useCellToString() returns (props) => string

The factory pattern supports conditional usage in consumer applications
where cellToString needs to be stored and called later based on runtime
conditions. Hook-based approach enables future CellContext integration
while maintaining the same functional API for renderer selection logic.

In order for linkCellToString functionality to leverage new
useCellToString hook, inline logic into LinkCell component. Since
linkCellToString was only used within LinkCell and returned JSX,
inlining eliminates the abstraction and keeps content rendering logic
properly encapsulated.
