+++ javascript/packages/core/components/cell/renderers/default-cell-renderer.tsx
@@ -1,8 +1,8 @@
 import { useStyletron } from 'baseui';
 
 import { cellTooltipHOC } from '#core/components/cell/components/tooltip/cell-tooltip-hoc';
-import { getCellRenderer } from '#core/components/cell/get-cell-renderer';
 import { useCellStyles } from '#core/components/cell/hooks';
+import { useGetCellRenderer } from '#core/components/cell/use-get-cell-renderer';
 
 import type { Cell, CellRendererProps } from '#core/components/cell/types';
 
@@ -11,6 +11,7 @@ export function DefaultCellRenderer(props: CellRendererProps<unknown, Cell>) {
   const { column, record } = props;
   const style = useCellStyles({ record, style: column.style });
 
+  const getCellRenderer = useGetCellRenderer();
   const ColumnRendererComponent = getCellRenderer(props);
   const Component = column.tooltip
     ? cellTooltipHOC(ColumnRendererComponent)
