+++ javascript/packages/core/components/cell/renderers/link/link-cell.tsx
@@ -1,25 +1,25 @@
 import { useStyletron } from 'baseui';
 
+import { useCellToString } from '#core/components/cell/use-cell-to-string';
 import { Icon } from '#core/components/icon/icon';
 import { Link } from '#core/components/link/link';
-import { linkCellToString } from './link-cell-to-string';
+import { TruncatedText } from '#core/components/truncated-text/truncated-text';
 
 import type { CellRendererProps } from '#core/components/cell/types';
 import type { LinkCellConfig } from './types';
 
 export function LinkCell(props: CellRendererProps<string, LinkCellConfig>) {
   const [css, theme] = useStyletron();
+  const cellToString = useCellToString();
   const { column } = props;
   const { icon, url } = column;
+
+  const content = <TruncatedText>{cellToString(props) ?? props.value}</TruncatedText>;
 
   return (
     <div className={css({ display: 'flex', alignItems: 'center', gap: theme.sizing.scale100 })}>
       {icon && <Icon name={icon} />}
-      {url && typeof url === 'string' ? (
-        <Link href={url}>{linkCellToString(props)}</Link>
-      ) : (
-        linkCellToString(props)
-      )}
+      {url && typeof url === 'string' ? <Link href={url}>{content}</Link> : content}
     </div>
   );
 }
