+++ javascript/packages/core/components/cell/renderers/text/text-cell.tsx
@@ -1,8 +1,9 @@
-import { cellToString } from '#core/components/cell/cell-to-string';
+import { useCellToString } from '#core/components/cell/use-cell-to-string';
 import { TruncatedText } from '#core/components/truncated-text/truncated-text';
 
 import type { CellRenderer } from '#core/components/cell/types';
 
 export const TextCell: CellRenderer<string> = (props) => {
+  const cellToString = useCellToString();
   return <TruncatedText>{cellToString(props) ?? '\u2014'}</TruncatedText>;
 };
